Conflicts:
	.travis.yml
	README.md
	SPREE_VERSION
	api/CHANGELOG.md
	api/app/controllers/spree/api/base_controller.rb
	api/app/controllers/spree/api/checkouts_controller.rb
	api/app/controllers/spree/api/products_controller.rb
	api/app/controllers/spree/api/shipments_controller.rb
	api/app/controllers/spree/api/stock_locations_controller.rb
	api/app/controllers/spree/api/taxons_controller.rb
	api/app/controllers/spree/api/variants_controller.rb
	api/app/models/spree/order_decorator.rb
	api/app/views/spree/api/products/index.v1.rabl
	api/config/routes.rb
	api/spec/controllers/spree/api/checkouts_controller_spec.rb
	api/spec/controllers/spree/api/orders_controller_spec.rb
	backend/CHANGELOG.md
	backend/app/assets/javascripts/admin/option_type_autocomplete.js.erb
	backend/app/assets/javascripts/admin/shipments.js.erb
	backend/app/assets/javascripts/admin/stock_transfer.js.coffee
	backend/app/assets/javascripts/admin/variant_autocomplete.js.erb
	backend/app/assets/stylesheets/admin/plugins/font-awesome.scss
	backend/app/controllers/spree/admin/products_controller.rb
	backend/app/controllers/spree/admin/stock_items_controller.rb
	backend/app/views/spree/admin/orders/_shipment.html.erb
	backend/app/views/spree/admin/orders/_shipment_manifest.html.erb
	backend/app/views/spree/admin/products/_form.html.erb
	backend/app/views/spree/admin/products/stock.html.erb
	backend/app/views/spree/admin/reports/index.html.erb
	backend/app/views/spree/admin/shared/_product_tabs.html.erb
	backend/app/views/spree/admin/shared/_routes.html.erb
	backend/app/views/spree/admin/shared/_translations.html.erb
	backend/app/views/spree/admin/shipping_methods/_form.html.erb
	backend/app/views/spree/admin/shipping_methods/index.html.erb
	backend/app/views/spree/admin/stock_locations/_form.html.erb
	backend/app/views/spree/admin/stock_locations/index.html.erb
	backend/app/views/spree/admin/taxons/_form.html.erb
	backend/app/views/spree/admin/variants/_autocomplete.js.erb
	backend/config/routes.rb
	backend/spec/features/admin/configuration/image_settings_spec.rb
	backend/spec/features/admin/configuration/stock_locations_spec.rb
	backend/spec/features/admin/orders/order_details_spec.rb
	backend/spec/features/admin/products/edit/products_spec.rb
	backend/spec/features/admin/products/option_types_spec.rb
	backend/spec/requests/admin/orders/payments_spec.rb
	backend/spec/requests/admin/orders/shipments_spec.rb
	backend/spec/requests/admin/products/stock_management_spec.rb
	backend/spree_backend.gemspec
	build.sh
	cmd/lib/spree_cmd/extension.rb
	cmd/lib/spree_cmd/installer.rb
	common_spree_dependencies.rb
	core/CHANGELOG.md
	core/app/controllers/spree/base_controller.rb
	core/app/mailers/spree/test_mailer.rb
	core/app/models/spree/address.rb
	core/app/models/spree/adjustment.rb
	core/app/models/spree/alert.rb
	core/app/models/spree/calculator/percent_per_item.rb
	core/app/models/spree/credit_card.rb
	core/app/models/spree/inventory_unit.rb
	core/app/models/spree/option_value.rb
	core/app/models/spree/order.rb
	core/app/models/spree/order/checkout.rb
	core/app/models/spree/order_contents.rb
	core/app/models/spree/order_inventory.rb
	core/app/models/spree/payment.rb
	core/app/models/spree/payment/processing.rb
	core/app/models/spree/price.rb
	core/app/models/spree/product.rb
	core/app/models/spree/product_property.rb
	core/app/models/spree/promotion/rules/item_total.rb
	core/app/models/spree/shipment.rb
	core/app/models/spree/shipping_method.rb
	core/app/models/spree/shipping_rate.rb
	core/app/models/spree/stock/availability_validator.rb
	core/app/models/spree/stock/estimator.rb
	core/app/models/spree/stock_item.rb
	core/app/models/spree/stock_location.rb
	core/app/models/spree/stock_movement.rb
	core/app/models/spree/taxon.rb
	core/app/models/spree/variant.rb
	core/app/models/spree/zone.rb
	core/config/locales/en.yml
	core/lib/spree/core.rb
	core/lib/spree/core/controller_helpers/order.rb
	core/lib/spree/core/engine.rb
	core/lib/spree/core/permalinks.rb
	core/lib/spree/core/version.rb
	core/lib/spree/testing_support/capybara_ext.rb
	core/lib/spree/testing_support/factories/product_factory.rb
	core/spec/models/spree/adjustment_spec.rb
	core/spec/models/spree/asset_spec.rb
	core/spec/models/spree/credit_card_spec.rb
	core/spec/models/spree/order_contents_spec.rb
	core/spec/models/spree/order_populator_spec.rb
	core/spec/models/spree/order_spec.rb
	core/spec/models/spree/order_updater_spec.rb
	core/spec/models/spree/payment_spec.rb
	core/spec/models/spree/product_spec.rb
	core/spec/models/spree/promotion/actions/create_line_items_spec.rb
	core/spec/models/spree/shipment_spec.rb
	core/spec/models/spree/stock_item_spec.rb
	core/spec/models/spree/stock_location_spec.rb
	core/spec/models/spree/variant_spec.rb
	core/spec/spec_helper.rb
	core/spree_core.gemspec
	frontend/CHANGELOG.md
	frontend/app/assets/javascripts/store/checkout.js.coffee
	frontend/app/assets/javascripts/store/product.js.coffee
	frontend/app/controllers/spree/checkout_controller.rb
	frontend/app/controllers/spree/taxons_controller.rb
	frontend/spec/controllers/spree/checkout_controller_spec.rb
	frontend/spec/controllers/spree/orders_controller_spec.rb
	frontend/spec/requests/address_spec.rb
	frontend/spree_frontend.gemspec
	lib/sandbox.sh
	sample/Gemfile

+++ core/app/models/spree/order_inventory.rb
@@ -6,10 +6,13 @@ module Spree
       @order = order
     end
 
-    # Only verify inventory for completed orders
-    # as carts have inventory assigned via create_proposed_shipment methh
+    # Only verify inventory for completed orders (as orders in frontend checkout
+    # have inventory assigned via +order.create_proposed_shipment+) or when
+    # shipment is explicitly passed
     #
-    # or when shipment is explicitly passed
+    # In case shipment is passed the stock location should only unstock or
+    # restock items if the order is completed. That is so because stock items
+    # are always unstocked when the order is completed through +shipment.finalize+
     def verify(line_item, shipment = nil)
       if order.completed? || shipment.present?
 
@@ -29,7 +32,7 @@ module Spree
     end
 
     def inventory_units_for(variant)
-      units = order.shipments.collect{|s| s.inventory_units.all}.flatten
+      units = order.shipments.collect{|s| s.inventory_units.to_a}.flatten
       units.group_by(&:variant_id)[variant.id] || []
     end
 
@@ -51,7 +54,6 @@ module Spree
     #
     # first unshipped that already includes this variant
     # first unshipped that's leaving from a stock_location that stocks this variant
-    #
     def determine_target_shipment(variant)
       shipment = order.shipments.detect do |shipment|
         (shipment.ready? || shipment.pending?) && shipment.include?(variant)
@@ -63,24 +65,16 @@ module Spree
     end
 
     def add_to_shipment(shipment, variant, quantity)
-      #create inventory_units
       on_hand, back_order = shipment.stock_location.fill_status(variant, quantity)
 
-      on_hand.times do
-        shipment.inventory_units.create({variant_id: variant.id,
-                                          state: 'on_hand'}, without_protection: true)
-      end
-
-      back_order.times do
-        shipment.inventory_units.create({variant_id: variant.id,
-                                         state: 'backordered'}, without_protection: true)
-      end
-
+      on_hand.times { shipment.set_up_inventory('on_hand', variant, order) }
+      back_order.times { shipment.set_up_inventory('backordered', variant, order) }
 
       # adding to this shipment, and removing from stock_location
-      shipment.stock_location.unstock variant, quantity, shipment
+      if order.completed?
+        shipment.stock_location.unstock(variant, quantity, shipment)
+      end
 
-      # return quantity added
       quantity
     end
 
@@ -99,14 +93,13 @@ module Spree
         removed_quantity += 1
       end
 
-      if shipment.inventory_units.count == 0
-        shipment.destroy
-      end
+      shipment.destroy if shipment.inventory_units.count == 0
 
       # removing this from shipment, and adding to stock_location
-      shipment.stock_location.restock variant, removed_quantity, shipment
+      if order.completed?
+        shipment.stock_location.restock variant, removed_quantity, shipment
+      end
 
-      # return quantity removed
       removed_quantity
     end
   end
